[Rails51] Base adapter tests passing.
* Fix #exec_insert to super up.
* Dup frozen strings passed into database statements.
* Remove tables/table_exists? deprecations.
* Switch to bigint primary keys.
* The `drop_table` with force cascade option now mimics in via pure SQL for us.
* Support MismatchedForeignKey exception.

+++ test/schema/sqlserver_specific_schema.rb
@@ -148,7 +148,7 @@ ActiveRecord::Schema.define do
 
   # Constraints
 
-  create_table(:sst_has_fks, force: true) { |t| t.column(:fk_id, :integer, null: false) }
+  create_table(:sst_has_fks, force: true) { |t| t.column(:fk_id, :bigint, null: false) }
   create_table(:sst_has_pks, force: true) { }
   execute <<-ADDFKSQL
     ALTER TABLE sst_has_fks
